Cleaned up nodeTree to be more nodeLink conscious

+++ src/objArray.h
@@ -286,7 +286,7 @@ extern "C" {
      */
     ERESULT         objArray_Put(
         OBJARRAY_DATA	*this,
-        uint16_t        index,          // Relative to 1
+        uint32_t        index,          // Relative to 1
         OBJ_ID          pObj
     );
